backend/requirements.txt
# Google ADK (Agent Development Kit)
google-adk

# Image processing dependencies
Pillow>=10.0.0

# Background removal
rembg[cpu]>=2.0.0

# Optional: For advanced image optimization
# pillow-avif-plugin>=1.3.0  # AVIF support
# pillow-heif>=0.13.0  # HEIF/HEIC support

# GenAI
google-genai

# AWS/S3
boto3
